+++ src/main/java/com/prgrms/needit/domain/user/favorite/repository/FavoriteCenterRepository.java
@@ -1,6 +1,6 @@
-package com.prgrms.needit.domain.user.member.repository;
+package com.prgrms.needit.domain.user.favorite.repository;
 
-import com.prgrms.needit.domain.user.member.entity.FavoriteCenter;
+import com.prgrms.needit.domain.user.favorite.entity.FavoriteCenter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FavoriteCenterRepository extends JpaRepository<FavoriteCenter, Long> {
