+++ src/main/java/com/prgrms/needit/domain/user/member/controller/MemberController.java
@@ -1,12 +1,10 @@
 package com.prgrms.needit.domain.user.member.controller;
 
 import com.prgrms.needit.common.response.ApiResponse;
-import com.prgrms.needit.domain.user.center.dto.CenterListResponse;
 import com.prgrms.needit.domain.user.member.dto.MemberCreateRequest;
 import com.prgrms.needit.domain.user.member.dto.MemberResponse;
 import com.prgrms.needit.domain.user.member.dto.MemberUpdateRequest;
 import com.prgrms.needit.domain.user.member.service.MemberService;
-import java.util.List;
 import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,11 +23,13 @@ public class MemberController {
 
 	private final MemberService memberService;
 
-	public MemberController(MemberService memberService) {
+	public MemberController(
+		MemberService memberService
+	) {
 		this.memberService = memberService;
 	}
 
-	@PostMapping("/signUp")
+	@PostMapping("/signup")
 	public ResponseEntity<ApiResponse<Long>> createMember(
 		@RequestBody @Valid MemberCreateRequest request
 	) {
@@ -61,18 +61,4 @@ public class MemberController {
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
 
-	@GetMapping("/favorites")
-	public ResponseEntity<ApiResponse<List<CenterListResponse>>> getFavCenters() {
-		return ResponseEntity.ok(
-			ApiResponse.of(memberService.getFavCenters())
-		);
-	}
-
-	@DeleteMapping("/favorites/{centerId}")
-	public ResponseEntity<ApiResponse<Void>> removeFavoriteCenter(
-		@PathVariable Long centerId
-	) {
-		memberService.removeFavoriteCenter(centerId);
-		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-	}
 }
