+++ src/main/java/com/prgrms/needit/domain/user/member/service/MemberService.java
@@ -2,18 +2,12 @@ package com.prgrms.needit.domain.user.member.service;
 
 import com.prgrms.needit.common.error.ErrorCode;
 import com.prgrms.needit.common.error.exception.NotFoundResourceException;
-import com.prgrms.needit.domain.user.center.dto.CenterListResponse;
-import com.prgrms.needit.domain.user.center.entity.Center;
-import com.prgrms.needit.domain.user.center.repository.CenterRepository;
-import com.prgrms.needit.domain.user.login.service.UserService;
 import com.prgrms.needit.domain.user.member.dto.MemberCreateRequest;
 import com.prgrms.needit.domain.user.member.dto.MemberResponse;
 import com.prgrms.needit.domain.user.member.dto.MemberUpdateRequest;
 import com.prgrms.needit.domain.user.member.entity.Member;
-import com.prgrms.needit.domain.user.member.repository.FavoriteCenterRepository;
 import com.prgrms.needit.domain.user.member.repository.MemberRepository;
-import java.util.List;
-import java.util.stream.Collectors;
+import com.prgrms.needit.domain.user.user.service.UserService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,21 +16,15 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberService {
 
 	private final MemberRepository memberRepository;
-	private final CenterRepository centerRepository;
-	private final FavoriteCenterRepository favoriteCenterRepository;
 	private final PasswordEncoder passwordEncoder;
 	private final UserService userService;
 
 	public MemberService(
 		MemberRepository memberRepository,
-		CenterRepository centerRepository,
-		FavoriteCenterRepository favoriteCenterRepository,
 		PasswordEncoder passwordEncoder,
 		UserService userService
 	) {
 		this.memberRepository = memberRepository;
-		this.centerRepository = centerRepository;
-		this.favoriteCenterRepository = favoriteCenterRepository;
 		this.passwordEncoder = passwordEncoder;
 		this.userService = userService;
 	}
@@ -80,40 +68,11 @@ public class MemberService {
 		curMember.deleteEntity();
 	}
 
-	@Transactional
-	public List<CenterListResponse> getFavCenters() {
-		Member curMember = userService.getCurMember()
-									  .orElseThrow();
-		return curMember.getFavoriteCenters()
-						.stream()
-						.map(favCenter -> new CenterListResponse(favCenter.getCenter()))
-						.collect(Collectors.toList());
-	}
-
-	@Transactional
-	public void removeFavoriteCenter(Long centerId) {
-
-		Member member = userService.getCurMember()
-								   .orElseThrow();
-		Center center = findActiveCenter(centerId);
-
-		member.deleteFavCenter(center);
-
-		favoriteCenterRepository.deleteByCenter(centerId);
-	}
-
 	public Member findActiveMember(Long memberId) {
 		return memberRepository
 			.findByIdAndIsDeletedFalse(memberId)
 			.orElseThrow(
 				() -> new NotFoundResourceException(ErrorCode.NOT_FOUND_MEMBER));
-	}
-
-	public Center findActiveCenter(Long centerId) {
-		return centerRepository
-			.findByIdAndIsDeletedFalse(centerId)
-			.orElseThrow(
-				() -> new NotFoundResourceException(ErrorCode.NOT_FOUND_CENTER));
 	}
 
 }
